create authMiddleware and move logic from route files to controllers file
