+++ backend/middleware/authMiddleware.ts
@@ -1,0 +1,28 @@
+import { Request, Response, NextFunction } from "express";
+import jwt from "jsonwebtoken";
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const secretKey = process.env.JWT_SECRET;
+
+if (!secretKey) {
+  throw new Error('JWT_SECRET is not defined in the environment variables');
+}
+
+const authMiddleware = (req: Request, res: Response, next: NextFunction) => {
+  const token = req.header('Authorization')?.replace('Bearer', '')
+
+  if (!token) {
+    return res.status(401).json({ message: 'Access denied. No token provided.' });
+  }
+
+  try {
+    const decoded = jwt.verify(token, secretKey)
+  }
+  catch (error) {
+    res.status(400).json({ message: 'Invalid token' });
+  }
+};
+
+export { authMiddleware };
