+++ backend/package.json
@@ -21,6 +21,7 @@
     "path": "^0.12.7"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/bcryptjs": "^2.4.6",
     "@types/cors": "^2.8.14",
     "@types/express": "^4.17.18",
